# openapi/pulse-api.yaml
openapi: 3.0.0
paths:
...
  /admin/tracks:
    get:
      description: Incluye las pistas en espera, que no aparecen en `/catalog`, y los álbumes a los que se pueden mover.
      operationId: listAdminTracks
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCatalogDto"
        "401": &a5
          description: Sin sesión válida ni token de servicio.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
        "403": &a6
          description: La sesión es válida pero la cuenta no es ADMIN.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
      security: &a7
        - AdminToken: []
        - BearerAuth: []
      summary: Listar el catálogo completo para mantenimiento
      tags: &a8
        - Admin
    post:
      description: "Única puerta de entrada al catálogo. Es idempotente por slug: reenviar una pista ya publicada la actualiza en vez de duplicarla, así que una subida cortada se puede reintentar tal cual. Lo que llega incompleto entra pero queda en espera, y se publica solo cuando no falta nada (ver `pending`)."
      operationId: publishTrack
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/PublishTrackFormDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishTrackResultDto"
        "400":
          description: Metadatos inválidos, o el audio no es un MP3 válido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
        "401": *a5
        "403": *a6
        "409":
          description: Ese mismo audio, byte a byte, ya está publicado con otro slug.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
        "413":
          description: El audio pasa de 60 MB o la portada de 5 MB.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
      security: *a7
      summary: Publicar una pista con su audio
      tags: *a8
  /admin/tracks/{identifier}:
    patch:
      operationId: updateAdminTrack
      parameters:
        - name: identifier
          required: true
          in: path
          description: UUID de la pista, su slug canónico, o un identificador antiguo `tr-*`.
          schema:
            example: get-lucky
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTrackDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminTrackDto"
        "400":
          description: Metadatos inválidos, o el álbum destino no existe.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
        "401": *a5
        "403": *a6
        "404":
          description: No existe ninguna pista con ese identificador.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
      security: *a7
      summary: Editar los metadatos de una pista
      tags: *a8
  /admin/albums/{id}:
    patch:
      description: El año es el único dato que retiene pistas fuera del catálogo. Al corregirlo se publican las que solo esperaban eso; una pista retirada a mano no vuelve.
      operationId: updateAdminAlbum
      parameters:
        - name: id
          required: true
          in: path
          description: UUID del álbum.
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAlbumDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateAlbumResultDto"
        "400":
          description: Datos inválidos o cuerpo sin nada que cambiar.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
        "401": *a5
        "403": *a6
        "404":
          description: No existe ningún álbum con ese UUID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
      security: *a7
      summary: Editar un álbum y rescatar sus pistas en espera
      tags: *a8
  /admin/artists/{id}:
    patch:
      description: "Géneros, reseña y retrato. El primer género es el que hereda cada pista suya que se publique sin género, así que corregirlo aquí evita que sigan cayendo en «Sin clasificar». Una publicación nunca pisa estos datos: este es el único sitio donde se cambian."
      operationId: updateAdminArtist
      parameters:
        - name: id
          required: true
          in: path
          description: UUID del artista.
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateArtistDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminArtistDto"
        "400":
          description: Datos inválidos o cuerpo sin nada que cambiar.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
        "401": *a5
        "403": *a6
        "404":
          description: No existe ningún artista con ese UUID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
      security: *a7
      summary: Editar la ficha de un artista
      tags: *a8
  /admin/tracks/{identifier}/audio:
    post:
      description: El archivo se guarda por el hash de su contenido, así que el `streamUrl` que devuelve trae una versión nueva y los clientes no reutilizan el MP3 anterior.
      operationId: replaceAdminTrackAudio
      parameters:
        - name: identifier
          required: true
          in: path
          description: UUID de la pista, su slug canónico, o un identificador antiguo `tr-*`.
          schema:
            example: get-lucky
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ReplaceAudioFormDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminTrackDto"
        "400":
          description: El archivo no es un MP3 válido, o falta la duración.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
        "401": *a5
        "403": *a6
        "404":
          description: No existe ninguna pista con ese identificador.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
        "413":
          description: El audio pasa de 60 MB.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
      security: *a7
      summary: Reemplazar el audio de una pista ya publicada
      tags: *a8
  /admin/covers:
    post:
      description: Devuelve la ruta ya servible. Se guarda por el hash del contenido, así que subir dos veces la misma imagen no ocupa el doble.
      operationId: uploadAdminCover
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UploadCoverFormDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoverUploadResultDto"
        "400":
          description: El archivo no es un JPG, PNG o WebP válido.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
        "401": *a5
        "403": *a6
        "413":
          description: La imagen pasa de 5 MB.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorDto"
      security: *a7
      summary: Subir una portada
      tags: *a8
info:
  title: Pulse Music API
  description: |-
    Catálogo y reproducción de la biblioteca de Pulse.

    La autenticación es global: salvo lo marcado como público —`/health`, `/catalog`,
    `/tracks/{id}/stream`, `/media/covers/{file}` y el propio `/auth`—, toda ruta exige sesión.

    `/tracks/{id}/stream` es público en el sentido de que no pide un Bearer: un `<audio src>` no
    puede mandar cabeceras. En su lugar exige la firma que emite `/catalog` en cada `streamUrl`,
    válida 24 horas. Sin ella responde 403, y la forma de conseguir una nueva es volver a pedir
    el catálogo.

    Lo que hay bajo `/admin` mantiene el catálogo y pide rol `ADMIN`. Se entra de dos maneras:
    con la sesión de una cuenta administradora, o con la cabecera `X-Pulse-Admin-Token`, que
    queda reservada a los procesos automatizados. Un credencial válido pero sin rol recibe 403,
    no 401: no hay nada que renovar.

    Lo que hay bajo `/auth` abre y renueva sesiones. El access token es un JWT de vida corta que
    viaja en `Authorization: Bearer`; el refresh token dura semanas y sale en una cookie
    `httpOnly` que solo se manda a `/v1/auth`. Un cliente sin cookies —la app móvil— lo pide en
    el cuerpo con la cabecera `X-Pulse-Client: native`.

    Toda ruta cuelga del prefijo `/v1`, ya incluido en la URL del servidor.
  version: 0.1.0
  contact: {}
...
servers:
  - url: http://localhost:4000/v1
    description: Servidor
components:
  securitySchemes:
    AdminToken:
      type: apiKey
      in: header
      name: X-Pulse-Admin-Token
      description: "Credencial de servicio para procesos sin persona detrás, igual a `PULSE_ADMIN_TOKEN` en el servidor. No es para el navegador: desde la web se entra con una cuenta de rol ADMIN."
    BearerAuth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: "Access token devuelto por `/auth/login`. Caduca a los 15 minutos: renuévalo en `/auth/refresh`."
  schemas:
...
    AudioCodec:
      type: string
      enum:
        - MP3
        - AAC
        - FLAC
...
    ApiErrorDto:
      type: object
      properties:
        statusCode:
          type: number
          example: 400
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          example:
            - title must be longer than or equal to 1 characters
        error:
          type: string
          example: Bad Request
      required:
        - statusCode
        - message
        - error
...
    AdminTrackDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
          example: get-lucky
        title:
          type: string
          example: Get Lucky
        artistId:
          type: string
          format: uuid
        artistName:
          type: string
          example: Daft Punk
        albumId:
          type: string
          format: uuid
        albumTitle:
          type: string
          example: Random Access Memories
        duration:
          type: number
          example: 369
          description: Duración en segundos.
        genre:
          type: string
          example: Electronic
        explicit:
          type: boolean
        codec:
          allOf:
            - $ref: "#/components/schemas/AudioCodec"
        coverUrl:
          type: string
          description: "Portada efectiva: la propia si existe, si no la del álbum."
        ownCoverUrl:
          type: string
          nullable: true
          description: Portada propia de la pista. `null` significa que hereda la del álbum.
        albumCoverUrl:
          type: string
        streamUrl:
          type: string
        fileSizeBytes:
          type: number
          example: 8412160
        isPublished:
          type: boolean
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - slug
        - title
        - artistId
        - artistName
        - albumId
        - albumTitle
        - duration
        - genre
        - explicit
        - codec
        - coverUrl
        - ownCoverUrl
        - albumCoverUrl
        - streamUrl
        - fileSizeBytes
        - isPublished
        - updatedAt
    AdminAlbumDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
          example: random-access-memories
        title:
          type: string
          example: Random Access Memories
        artistId:
          type: string
          format: uuid
        artistName:
          type: string
          example: Daft Punk
        year:
          type: number
          example: 2013
          description: "`0` marca un álbum sin año, que retiene sus pistas en espera."
        coverUrl:
          type: string
      required:
        - id
        - slug
        - title
        - artistId
        - artistName
        - year
        - coverUrl
    AdminArtistDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
          example: daft-punk
        name:
          type: string
          example: Daft Punk
        photoUrl:
          type: string
          nullable: true
          description: "`null` cuando el artista no tiene retrato. A diferencia del catálogo público, aquí no se rellena con el placeholder: quien vaya a copiar la foto necesita distinguir un retrato real del relleno que llevan todos."
        genres:
          example:
            - Reggae
            - Dancehall
          description: "Géneros curados del artista, del más general al más específico. Los publica este endpoint porque quien sube pistas los necesita para no mandarlas sin género: sin ellos aquí, un cliente solo podía enviar una lista vacía y toda pista nueva caía en «Sin clasificar»."
          type: array
          items:
            type: string
        bio:
          type: string
          nullable: true
          description: Reseña del artista, si la tiene.
      required:
        - id
        - slug
        - name
        - photoUrl
        - genres
        - bio
    AdminCatalogDto:
      type: object
      properties:
        tracks:
          description: Todas las pistas, publicadas o no.
          type: array
          items:
            $ref: "#/components/schemas/AdminTrackDto"
        albums:
          description: Álbumes válidos como destino de una pista.
          type: array
          items:
            $ref: "#/components/schemas/AdminAlbumDto"
        artists:
          description: Los artistas, con su foto si la tienen.
          type: array
          items:
            $ref: "#/components/schemas/AdminArtistDto"
      required:
        - tracks
        - albums
        - artists
    UpdateTrackDto:
      type: object
      properties:
        title:
          type: string
          maxLength: 120
          example: Jigsaw Falling Into Place
        genre:
          type: string
          maxLength: 80
          example: Alternative rock
        explicit:
          type: boolean
          example: false
        albumId:
          type: string
          format: uuid
          description: Mover la pista a otro álbum. El artista pasa a ser el del álbum destino.
        isPublished:
          type: boolean
          description: Publica o retira la pista del catálogo público.
        coverUrl:
          type: string
          nullable: true
          maxLength: 2048
          example: /media/covers/9f86d0….jpg
          description: URL HTTP(S) o ruta /media/covers/. `null` borra la portada propia y hereda la del álbum.
    UpdateAlbumDto:
      type: object
      properties:
        year:
          type: number
          minimum: 1900
          maximum: 2100
          example: 2013
          description: Corregirlo publica las pistas del álbum que solo esperaban este dato.
        title:
          type: string
          maxLength: 200
          example: Random Access Memories
        coverUrl:
          type: string
          maxLength: 2048
          example: /media/covers/9f86d0….jpg
          description: "Portada del disco. Útil para los recopilatorios que no tienen carátula propia: se les puede poner la foto del artista."
    UpdateAlbumResultDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
          example: random-access-memories
        title:
          type: string
          example: Random Access Memories
        year:
          type: number
          example: 2013
        coverUrl:
          type: string
        published:
          type: number
          example: 2
          description: Pistas que estaban en espera por este álbum y han pasado a publicadas.
      required:
        - id
        - slug
        - title
        - year
        - coverUrl
        - published
    UpdateArtistDto:
      type: object
      properties:
        genres:
          maxItems: 8
          example:
            - Reggae
            - Dancehall
          description: "Géneros del artista, del más general al más específico. El primero es el que hereda una pista suya publicada sin género, así que el orden no es decorativo. Sustituye la lista entera: mandar `[]` la deja vacía."
          type: array
          items:
            type: string
        bio:
          type: string
          nullable: true
          maxLength: 4000
          description: Reseña del artista. `null` la quita.
        photoUrl:
          type: string
          nullable: true
          maxLength: 2048
          example: /media/covers/9f86d0….jpg
          description: Retrato del artista. `null` lo quita y el catálogo público vuelve al relleno.
    ReplaceAudioFormDto:
      type: object
      properties:
        audio:
          type: string
          format: binary
          description: MP3, máximo 60 MB.
        duration:
          type: number
          example: 369
          description: Duración del nuevo audio, en segundos.
      required:
        - audio
        - duration
    UploadCoverFormDto:
      type: object
      properties:
        cover:
          type: string
          format: binary
          description: JPG, PNG o WebP, máximo 5 MB.
      required:
        - cover
    CoverUploadResultDto:
      type: object
      properties:
        coverUrl:
          type: string
          example: /media/covers/9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08.jpg
          description: Ruta servida por este API. Sirve tal cual como `coverUrl` al editar una pista.
      required:
        - coverUrl
...
    PublishTrackFormDto:
      type: object
      properties:
        audio:
          type: string
          format: binary
          description: MP3, máximo 60 MB.
        data:
          type: string
          description: PublishTrackDto serializado como JSON.
          example: '{"slug":"get-lucky","title":"Get Lucky","duration":369,"artist":{"slug":"daft-punk","name":"Daft Punk","genres":["Electronic"]},"album":{"slug":"random-access-memories","title":"Random Access Memories","year":2013},"genre":"Electronic"}'
        cover:
          type: string
          format: binary
          description: JPG, PNG o WebP, máximo 5 MB.
      required:
        - audio
        - data
    PublishTrackResultDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
          example: get-lucky
        title:
          type: string
          example: Get Lucky
        artistId:
          type: string
          format: uuid
        artistName:
          type: string
          example: Daft Punk
        albumId:
          type: string
          format: uuid
        albumTitle:
          type: string
          example: Random Access Memories
        duration:
          type: number
          example: 369
          description: Duración en segundos.
        genre:
          type: string
          example: Electronic
        explicit:
          type: boolean
        codec:
          allOf:
            - $ref: "#/components/schemas/AudioCodec"
        coverUrl:
          type: string
          description: "Portada efectiva: la propia si existe, si no la del álbum."
        ownCoverUrl:
          type: string
          nullable: true
          description: Portada propia de la pista. `null` significa que hereda la del álbum.
        albumCoverUrl:
          type: string
        streamUrl:
          type: string
        fileSizeBytes:
          type: number
          example: 8412160
        isPublished:
          type: boolean
        updatedAt:
          type: string
          format: date-time
        pending:
          example:
            - año
          description: Datos que faltan para publicar. Si no está vacío la pista queda en espera, fuera del catálogo público.
          type: array
          items:
            type: string
      required:
        - id
        - slug
        - title
        - artistId
        - artistName
        - albumId
        - albumTitle
        - duration
        - genre
        - explicit
        - codec
        - coverUrl
        - ownCoverUrl
        - albumCoverUrl
        - streamUrl
        - fileSizeBytes
        - isPublished
        - updatedAt
        - pending